searchable: follow SDK guidelines
Following: https://developer.android.com/guide/topics/search/searchable-config.html

android:label
String resource. (Required.) The name of your application. It should be the same as the name applied to the android:label attribute of your <activity> or <application> manifest element.
This label is only visible to the user when you set android:includeInGlobalSearch to "true", in which case, this label is used to identify your application as a searchable item in the system's search settings.

android:hint
String resource. (Recommended.) The text to display in the search text field when no text has been entered. It provides a hint to the user about what content is searchable.
For consistency with other Android applications, you should format the string for android:hint as "Search <content-or-product>". For example, "Search songs and artists" or "Search YouTube".